requirements.txt
langchain==0.3.13
langchain-community==0.3.13
pypdf==5.1.0
python-dotenv==1.0.1
sentence-transformers==3.3.1
pinecone-client==5.0.1
ctransformers==0.2.27
Flask==3.1.0
-e .
